Fi responsive ui
