+++ react-client/src/layouts/ExploreSection1.tsx
@@ -1,24 +1,45 @@
-/* eslint-disable no-nested-ternary */
+/* eslint-disable consistent-return */
+
 /* eslint-disable no-param-reassign */
 import SearchTrackBox from '@components/SearchTrackBox';
 import { Color, SelectedTrack } from '@pages/ExplorePage';
 import { TrackWithArtistResponse } from '@typings/track';
+import { useEffect, useRef, useState } from 'react';
 import { Updater } from 'use-immer';
 
 interface Prob{
     selectedTracks: SelectedTrack[]; // 상태 값
     setSelectedTracks: Updater<SelectedTrack[]>;
-    containerWidth:number
 }
 
 export default function ExploreSection1({
   setSelectedTracks,
-  selectedTracks, containerWidth,
+  selectedTracks,
 }:Prob) {
   const additionalBoxRenderCondition = !!(
     selectedTracks[selectedTracks.length - 1]?.track && selectedTracks.length < 6);
-  const totalBoxes = selectedTracks.length + (additionalBoxRenderCondition ? 1 : 0);
+  const additionalBoxRenderCondition2 = selectedTracks.length === 0;
+  const totalBoxes = selectedTracks.length + ((additionalBoxRenderCondition
+     || additionalBoxRenderCondition2) ? 1 : 0);
   const colorArray = Object.values(Color);
+  const [containerWidth, setContainerWidth] = useState<number>(0); // 뷰포트
+  const containerRef = useRef<HTMLInputElement>(null);
+  // 브라우저 리사이즈 감지
+  useEffect(() => {
+    const container = containerRef.current;
+    if (!container) return;
+
+    const resizeObserver = new ResizeObserver((entries) => {
+      const entry = entries[0];
+      const { width } = entry.contentRect;
+      setContainerWidth(width);
+    });
+
+    resizeObserver.observe(container);
+    return () => {
+      resizeObserver.unobserve(container);
+    };
+  }, [containerRef]);
 
   function selectTrack(id: number, selectedTrack: TrackWithArtistResponse) {
     setSelectedTracks((draft) => {
@@ -43,16 +64,6 @@ export default function ExploreSection1({
     });
   }
 
-  function calculateBoxWidth() {
-    if (containerWidth >= 1024) {
-      return totalBoxes === 1 ? '90%' : totalBoxes === 2 ? '45%' : '30%';
-    }
-    if (containerWidth >= 640) {
-      return totalBoxes === 1 ? '90%' : '45%';
-    }
-    return '100%';
-  }
-
   function addSelectBox(id:number) {
     setSelectedTracks((draft) => {
       draft.push({ id, activate: true, color: colorArray[id] });
@@ -68,9 +79,19 @@ export default function ExploreSection1({
       }
     });
   }
+  function calculateBoxWidth() {
+    const gap = 2 * (totalBoxes - 1);
+    if (containerWidth < 640) {
+      return '100%';
+    }
+    if (totalBoxes <= 3) {
+      return (containerWidth - gap * 8) / totalBoxes;
+    }
+    return (containerWidth - gap * 8) / 3;
+  }
 
   return (
-    <section className="flex flex-wrap gap-4 items-center justify-center mt-[5rem] w-[90%]">
+    <section ref={containerRef} className="flex flex-wrap gap-2 items-center justify-center mt-[5rem] w-[100%] md:w-[90%] lg:w-[80%]">
       {selectedTracks.map((selectedTrack) => (
         <div style={{ width: calculateBoxWidth() }}>
           <SearchTrackBox
@@ -84,7 +105,7 @@ export default function ExploreSection1({
       ))}
 
       {/* 마지막 인덱스에 추가 박스 렌더링 */}
-      {additionalBoxRenderCondition && (
+      {(additionalBoxRenderCondition || additionalBoxRenderCondition2) && (
         <div style={{ width: calculateBoxWidth() }}>
           <SearchTrackBox
             selectedTrack={{
@@ -96,7 +117,7 @@ export default function ExploreSection1({
             addSelectBox={addSelectBox}
             deleteSelectBox={deleteSelectBox}
             deleteTrack={deleteTrack}
-        />
+          />
         </div>
       )}
     </section>
