+++ react-client/src/layouts/ExploreSection2.tsx
@@ -46,50 +46,48 @@ export default function ExploreSection2({ selectedTracks }:Prob) {
   };
 
   return (
-    <section className="flex flex-wrap gap-4 items-center justify-center mt-[5rem] w-[90%] text-gray-700">
+    <section className="flex flex-wrap gap-2 items-center justify-center mt-[5rem] text-gray-700 w-[100%] md:w-[90%] lg:w-[80%]">
       {selectedTracks.map((selectedTrack) => {
         if (selectedTrack.track) {
           return (
-            <div key={selectedTrack.id} className='mb-8'>
-              <div className='text-lg mb-4'>{selectedTrack.track.titleName}</div>
+            <div key={selectedTrack.id} className="mb-8 w-full">
+              <div className="text-lg mb-4">{selectedTrack.track.titleName}</div>
 
-              <div className='flex gap-4'>
-                <div className='bg-white p-6 rounded-lg'>
-                  <div className='flex items-center mb-8 '>
-                    <div className='text-base px-2'>플랫폼별 차트순위</div>
+              {/* 부모 */}
+              <div className="flex gap-2 w-full">
+                {/* 자식1 */}
+                <div className="bg-white p-6 rounded-lg w-[50%] lg:w-[60%] ">
+                  <div className="flex items-center mb-8">
+                    <div className="text-base px-2">플랫폼별 차트순위</div>
                     <RxQuestionMarkCircled size={20} />
                   </div>
 
                   <TrackOverview key={selectedTrack.id} track={selectedTrack.track} />
                 </div>
 
-                <div className=''>
-                  <div className='bg-white p-6 rounded-lg mb-4'>
-                    <div className='px-2 mb-8'>아티스트</div>
+                {/* 자식2 */}
+                <div className="bg-white p-6 rounded-lg w-[50%] lg:w-[40%] ">
+                  <div className="px-2 mb-8">아티스트</div>
 
-                    {selectedTrack.track.artists.length > 1 ? (
-                      <div className='w-full'>
-                        <Slider {...settings} className="w-[31rem] ">
-                          {selectedTrack.track.artists.map((artist) => (
-                            <div className='border border-gray-300 '>
-                              <ArtistsInfoCard artist={artist} size={100} />
-                            </div>
-                          ))}
-                        </Slider>
-                      </div>
-                    ) : (
-                      <div className='border border-gray-300'>
-                        <ArtistsInfoCard artist={selectedTrack.track.artists[0]} size={100} />
-                      </div>
-                    )}
-
-                  </div>
+                  {selectedTrack.track.artists.length > 1 ? (
+                    <div className="w-full">
+                      <Slider {...settings} className="w-[31rem] ">
+                        {selectedTrack.track.artists.map((artist) => (
+                          <div className="border border-gray-300">
+                            <ArtistsInfoCard artist={artist} size={100} />
+                          </div>
+                        ))}
+                      </Slider>
+                    </div>
+                  ) : (
+                    <div className="border border-gray-300">
+                      <ArtistsInfoCard artist={selectedTrack.track.artists[0]} size={100} />
+                    </div>
+                  )}
 
                   <PlatformAnalysisBox platforms={selectedTrack.track.platforms} />
                 </div>
-
               </div>
-
             </div>
           );
         }
