+++ react-client/src/layouts/HomeSection2.tsx
@@ -148,7 +148,7 @@ export default function HomeSection2() {
 
         <h3 className="relative text-center text-[#444746] mb-12">플랫폼별로 서로 다른 차트 순위를 한눈에 비교해보세요.</h3>
 
-        <Slider {...settings} className="lg:w-[70rem] sm:w-[30rem] w-[22rem] ">
+        <Slider {...settings} className="xl:w-[80rem] lg:w-[60rem] sm:w-[30rem] w-[22rem] ">
           {contents}
         </Slider>
       </div>
