+++ react-client/src/layouts/TopNavBar.tsx
@@ -1,84 +1,108 @@
 import Logo from '@components/Logo';
 import { useModal } from '@hooks/useModal';
-import DefaultSearchBar from '@components/DefaultSearchBar';
+import React from 'react';
+import {
+  FiHome, FiSearch, FiTrendingUp, FiMessageSquare, FiMenu,
+} from 'react-icons/fi';
 import { Link } from 'react-router-dom';
-import DashboardSearchBar from '@components/DashboardSearchBar';
 
-interface prob {
-  page?: 'dashboard' | 'real-time';
-}
+type TopNavbarProps = {
+  currentPage: 'home'|'explore'|'realtime';
+};
 
-function TopNavBar({ page }: prob) {
-  const { isModalOpen, setIsModalOpen, modalRef } = useModal();
+function TopNavbar({ currentPage }: TopNavbarProps) {
+  const {
+    isModalOpen,
+    setIsModalOpen,
+    modalRef,
+  } = useModal();
+
+  function handleToggleButton(e: React.MouseEvent<HTMLButtonElement>) {
+    e.stopPropagation();
+    setIsModalOpen((prev) => !prev);
+  }
+
+  // 페이지 아래에 효과를 주는 함수
+  function getBorderClass(page: string) {
+    return currentPage === page ? 'absolute top-10 left-0 w-full h-1 bg-blue-500 rounded-t-full' : '';
+  }
 
   return (
-    <nav className="fixed z-50 top-0  left-0 w-full bg-white shadow-md">
-      <div className="container flex justify-center items-center h-16">
-        {/* Logo */}
-        <Logo />
-        {/* Center: Search Bar */}
-        {page === 'dashboard' ? <DashboardSearchBar /> : <DefaultSearchBar />}
+    <div className='relative z-10'>
+      <nav className="fixed top-0 right-0 z-10 w-full bg-white border-b border-gray-200">
+        <div className="z-10 flex items-center justify-between px-4 pr-8 py-2">
+          <div className="flex items-center">
+            <Logo />
 
-        {/* Right: Buttons for large screens */}
-        <div className="hidden md:flex items-center space-x-4">
-          <Link to={{ pathname: '/dashboard' }}>
-            <button
-            className={`px-[0.5rem] responsive-text font-bold rounded hover:text-[#9ab867] whitespace-nowrap ${
-              page === 'dashboard' ? 'text-[#9ab867]' : ''
-            }`}
-          >
-              DashBoard
+            <div className="hidden md:flex space-x-14 ml-8 text-gray-500">
+              <div className="relative">
+                <Link to={'/'} >
+                  홈
+                </Link>
+
+                <div className={getBorderClass('home')} />
+              </div>
+
+              <div className="relative">
+                <Link to={'/explore'}>
+                  탐색
+                </Link>
+
+                <div className={getBorderClass('explore')} />
+              </div>
+
+              <div className="relative">
+                <Link to={'/realtime'}>
+                  실시간
+                </Link>
+
+                <div className={getBorderClass('realtime')} />
+              </div>
+            </div>
+          </div>
+
+          <div>
+            <button onClick={handleToggleButton} className='hover:bg-gray-300 rounded-full p-2'>
+              <FiMenu className="w-6 h-6 text-gray-700" />
             </button>
+          </div>
+        </div>
+      </nav>
+
+      {/* 모달에 애니메이션 추가 */}
+      <nav
+      ref={modalRef}
+        className={`fixed top-0 right-0 z-5 h-full bg-white shadow-md w-[20rem] transition-transform transform duration-300 ${
+          isModalOpen ? 'translate-x-0' : 'translate-x-full'
+        }`}
+      >
+        <div className="flex flex-col  p-2 mt-16 text-gray-500">
+          <Link to="/" className={`flex items-center p-5 rounded-full ${currentPage === 'home' ? 'bg-[#c2e7ff]' : 'hover:bg-gray-100'}`}>
+            <FiHome className="mr-4 w-5 h-5" />
+            홈
           </Link>
 
-          <span className="text-gray-300">|</span>
+          <Link to="/explore" className={`flex items-center p-5 rounded-full ${currentPage === 'explore' ? 'bg-[#c2e7ff]' : 'hover:bg-gray-100'}`}>
+            <FiSearch className="mr-4 w-5 h-5" />
+            탐색
+          </Link>
 
-          <button
-            className={`px-[0.5rem] responsive-text font-bold rounded hover:text-[#9ab867] whitespace-nowrap ${
-              page === 'real-time' ? 'text-[#9ab867]' : ''
-            }`}
-          >
-            Real-Time Chart
-          </button>
+          <Link to="/realtime" className={`flex items-center p-5 rounded-full ${currentPage === 'realtime' ? 'bg-[#c2e7ff]' : 'hover:bg-gray-100'}`}>
+            <FiTrendingUp className="mr-4 w-5 h-5" />
+            실시간 인기
+          </Link>
+
+          <hr className="border-gray-300 my-4" />
+
+          {/* 의견 보내기 */}
+          <Link to="/feedback" className="flex items-center p-5 rounded-full hover:bg-gray-100">
+            <FiMessageSquare className="mr-4 w-5 h-5" />
+            의견 보내기
+          </Link>
         </div>
-
-        {/* Hamburger Menu for small screens */}
-        <div className="md:hidden flex items-center">
-          <button
-            className="px-[0.5rem] font-bold rounded focus:outline-none"
-            onClick={() => setIsModalOpen(!isModalOpen)}
-          >
-            &#x22EE;
-          </button>
-
-          {isModalOpen && (
-            <div
-              ref={modalRef}
-              className="absolute right-0 mt-2 w-48 bg-white shadow-lg rounded-lg"
-            >
-              <Link to={{ pathname: '/dashboard' }}>
-                <button
-                className={`block w-full text-left px-4 py-2 font-bold hover:text-[#9ab867] ${
-                  page === 'dashboard' ? 'text-[#9ab867]' : ''
-                }`}
-              >
-                  DashBoard
-                </button>
-              </Link>
-
-              <button
-                className={`block w-full text-left px-4 py-2 font-bold hover:text-[#9ab867] ${
-                  page === 'real-time' ? 'text-[#9ab867]' : ''
-                }`}
-              >
-                Real-Time Chart
-              </button>
-            </div>
-          )}
-        </div>
-      </div>
-    </nav>
+      </nav>
+    </div>
   );
 }
 
-export default TopNavBar;
+export default TopNavbar;
